Upgrade in-tree TAs to use latest version of TEE Internal Core API
Upgrades in-tree TAs to use latest version of TEE Internal Core API and
the compatibility flag is by default turned off.

Upgrading the TAs is basically a matter of replacing a few uint32_t
pointers with size_t pointers instead. This is done in as few places as
possible. So in particular the PKCS#11 TA might benefit from another
round of refactoring now that the contagious use of uint32_t as size
field is removed.

+++ ta/avb/entry.c
@@ -78,7 +78,7 @@ static TEE_Result read_rb_idx(uint32_t pt, TEE_Param params[TEE_NUM_PARAMS])
 						TEE_PARAM_TYPE_NONE);
 	size_t slot_offset;
 	uint64_t idx;
-	uint32_t count;
+	size_t count;
 	TEE_Result res;
 	TEE_ObjectHandle h;
 
@@ -134,7 +134,7 @@ static TEE_Result write_rb_idx(uint32_t pt, TEE_Param params[TEE_NUM_PARAMS])
 	size_t slot_offset;
 	uint64_t widx;
 	uint64_t idx;
-	uint32_t count;
+	size_t count;
 	TEE_Result res;
 	TEE_ObjectHandle h;
 
@@ -182,7 +182,7 @@ static TEE_Result read_lock_state(uint32_t pt, TEE_Param params[TEE_NUM_PARAMS])
 						TEE_PARAM_TYPE_NONE,
 						TEE_PARAM_TYPE_NONE);
 	uint32_t lock_state;
-	uint32_t count;
+	size_t count;
 	TEE_Result res;
 	TEE_ObjectHandle h;
 
@@ -220,7 +220,7 @@ static TEE_Result write_lock_state(uint32_t pt,
 						TEE_PARAM_TYPE_NONE);
 	uint32_t wlock_state;
 	uint32_t lock_state;
-	uint32_t count;
+	size_t count;
 	TEE_Result res;
 	TEE_ObjectHandle h;
 
@@ -316,7 +316,7 @@ static TEE_Result read_persist_value(uint32_t pt,
 	char *name_buf = NULL;
 	uint32_t value_sz = 0;
 	char *value = NULL;
-	uint32_t count = 0;
+	size_t count = 0;
 
 	if (pt != exp_pt)
 		return TEE_ERROR_BAD_PARAMETERS;
